Add exception handling closes #10

+++ src/main/java/com/voltaire/restaurant/MenuItemService.java
@@ -1,8 +1,10 @@
 package com.voltaire.restaurant;
 
+import com.voltaire.exception.customexceptions.EntityNotFoundException;
 import com.voltaire.restaurant.model.MenuItem;
 import com.voltaire.restaurant.model.MenuItemDto;
 import com.voltaire.restaurant.repository.MenuItemRepository;
+import com.voltaire.shared.IdResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -26,25 +28,35 @@ public class MenuItemService {
         return menuItemRepository.save(menuItem);
     }
 
+    public boolean notExists(Long id) {
+        return !menuItemRepository.existsById(id);
+    }
+
+    public IdResponse updateMenuItem(Long id, MenuItem menuItem) {
+        if (notExists(id)) {
+            throw new EntityNotFoundException(MenuItem.class, "id", id.toString());
+        }
+
+        menuItem.setId(id);
+        return new IdResponse(menuItemRepository.save(menuItem).getId());
+    }
+
+    public IdResponse deleteMenuItem(Long id) {
+        if (notExists(id)) {
+            throw new EntityNotFoundException(MenuItem.class, "id", id.toString());
+        }
+
+        menuItemRepository.deleteById(id);
+        return new IdResponse(id);
+    }
+
     public List<MenuItem> findAll() {
         return menuItemRepository.findAll();
     }
 
     public MenuItem findById(Long id) {
-        return menuItemRepository.findById(id).orElse(null);
-    }
-
-    public boolean notExists(Long id) {
-        return !menuItemRepository.existsById(id);
-    }
-
-    public void updateMenuItem(Long id, MenuItem menuItem) {
-        menuItem.setId(id);
-        menuItemRepository.save(menuItem);
-    }
-
-    public void deleteMenuItem(Long id) {
-        menuItemRepository.deleteById(id);
+        return menuItemRepository.findById(id).orElseThrow(() ->
+                new EntityNotFoundException(MenuItem.class, "id", id.toString()));
     }
 
 }
