+++ src/main/java/com/voltaire/restaurant/RestaurantService.java
@@ -1,7 +1,9 @@
 package com.voltaire.restaurant;
 
+import com.voltaire.exception.customexceptions.EntityNotFoundException;
 import com.voltaire.restaurant.model.Restaurant;
 import com.voltaire.restaurant.repository.RestaurantRepository;
+import com.voltaire.shared.IdResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -17,25 +19,34 @@ public class RestaurantService {
         return restaurantRepository.save(restaurant);
     }
 
+    public boolean notExists(Long id) {
+        return !restaurantRepository.existsById(id);
+    }
+
+    public IdResponse updateRestaurant(Long id, Restaurant restaurant) {
+        if (notExists(id)) {
+            throw new EntityNotFoundException(Restaurant.class, "id", id.toString());
+        }
+
+        restaurant.setId(id);
+        return new IdResponse(restaurantRepository.save(restaurant).getId());
+    }
+
+    public IdResponse deleteRestaurant(Long id) {
+        if (notExists(id)) {
+            throw new EntityNotFoundException(Restaurant.class, "id", id.toString());
+        }
+        restaurantRepository.deleteById(id);
+        return new IdResponse(id);
+    }
+
     public List<Restaurant> findAll() {
         return restaurantRepository.findAll();
     }
 
     public Restaurant findById(Long id) {
-        return restaurantRepository.findById(id).orElse(null);
-    }
-
-    public boolean notExists(Long id) {
-        return !restaurantRepository.existsById(id);
-    }
-
-    public void updateRestaurant(Long id, Restaurant restaurant) {
-        restaurant.setId(id);
-        restaurantRepository.save(restaurant);
-    }
-
-    public void deleteRestaurant(Long id) {
-        restaurantRepository.deleteById(id);
+        return restaurantRepository.findById(id).orElseThrow(() ->
+                new EntityNotFoundException(Restaurant.class, "id", id.toString()));
     }
 
 }
